+++ tests/app/Http/Requests/Filter/FilterCollectionTest.php
@@ -1,0 +1,47 @@
+<?php
+
+/**
+ * Copyright (c) 2018-2023 Adshares sp. z o.o.
+ *
+ * This file is part of AdServer
+ *
+ * AdServer is free software: you can redistribute and/or modify it
+ * under the terms of the GNU General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * AdServer is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty
+ * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with AdServer. If not, see <https://www.gnu.org/licenses/>
+ */
+
+namespace Adshares\Adserver\Tests\Http\Requests\Filter;
+
+use Adshares\Adserver\Http\Requests\Filter\FilterCollection;
+use Adshares\Adserver\Http\Requests\Filter\FilterType;
+use Adshares\Adserver\Http\Requests\Filter\StringFilter;
+use Adshares\Adserver\Tests\TestCase;
+use Illuminate\Http\Request;
+
+final class FilterCollectionTest extends TestCase
+{
+    public function testFromRequestStringZero(): void
+    {
+        $request = new Request(['filter' => ['query' => '0']]);
+
+        $filters = FilterCollection::fromRequest($request, [
+            'query' => FilterType::String,
+        ]);
+
+        $queryFilter = $filters->getFilterByName('query');
+        self::assertInstanceOf(StringFilter::class, $queryFilter);
+        /** @var StringFilter $queryFilter */
+        $values = $queryFilter->getValues();
+        self::assertCount(1, $values);
+        self::assertEquals('0', $values[0]);
+    }
+}