+++ CHANGELOG.md
@@ -7,6 +7,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Dynamic find with empty types or mimes
+
+### Fixed
+- Banner classification list (correct medium)
+
+## [2.3.2] - 2023-01-31
+### Fixed
+- Calculating waiting payments
+
+## [2.3.1] - 2023-01-31
+### Fixed
+- Acceptance for active sites
+- Account confirmation in case of registration during find
+- Filtering by non-empty string which casts to false, e.g. '0'
 
 ## [2.3.0] - 2023-01-26
 ### Added
@@ -818,7 +833,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v2.3.0...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v2.3.2...develop
+[2.3.2]: https://github.com/adshares/adserver/compare/v2.3.1...v2.3.2
+[2.3.1]: https://github.com/adshares/adserver/compare/v2.3.0...v2.3.1
 [2.3.0]: https://github.com/adshares/adserver/compare/v2.2.7...v2.3.0
 [2.2.7]: https://github.com/adshares/adserver/compare/v2.2.6...v2.2.7
 [2.2.6]: https://github.com/adshares/adserver/compare/v2.2.5...v2.2.6